refactor(tests): consolidate _FakeConnector into tests/helpers/fake_connector.py (#5558)
Extracts two inline _FakeConnector implementations per #5443/#5558:
- tests/helpers/fake_connector.py: single FakeConnector(config, result, delay=0) covering both use cases
- knowledge/connectors/registry_health_test.py: _FakeConnector → FakeConnector (imported)
- api/knowledge_connectors_health_test.py: _FakeConnector → FakeConnector (imported, delay defaults to 0)
